Refactored user controller
