+++ src/Tooling/Bit.Tooling.Templates/TodoTemplate/Api/Controllers/UserController.cs
@@ -9,18 +9,12 @@ public class UserController : ControllerBase
 {
     private readonly UserManager<User> _userManager;
 
-    private readonly IJwtService _jwtService;
-
     private readonly IMapper _mapper;
 
-    private readonly SignInManager<User> _signInManager;
-
-    public UserController(SignInManager<User> signInManager, UserManager<User> userManager, IJwtService jwtService, IMapper mapper)
+    public UserController(UserManager<User> userManager, IMapper mapper)
     {
         _userManager = userManager;
-        _jwtService = jwtService;
         _mapper = mapper;
-        _signInManager = signInManager;
     }
 
     [HttpGet("[action]")]
@@ -29,7 +23,7 @@ public class UserController : ControllerBase
         var userId = User.GetUserId();
 
         var user = await _userManager.Users.FirstOrDefaultAsync(user => user.Id == userId, cancellationToken);
-        
+
         if (user is null)
             throw new ResourceNotFoundException();
 
@@ -49,35 +43,5 @@ public class UserController : ControllerBase
         var updatedUser = _mapper.Map(userDto, user);
 
         await _userManager.UpdateAsync(updatedUser);
-    }
-
-    [HttpPost("[action]"), AllowAnonymous]
-    public async Task SignUp(UserDto dto)
-    {
-        var userToAdd = _mapper.Map<User>(dto);
-
-        var result = await _userManager.CreateAsync(userToAdd, dto.Password);
-
-        if (!result.Succeeded)
-            throw new ResourceValidationException(result.Errors.Select(e => e.Code).ToArray());
-    }
-
-    [HttpPost("[action]"), AllowAnonymous]
-    public async Task<SignInResponseDto> SignIn(SignInRequestDto requestToken)
-    {
-        var user = await _userManager.FindByNameAsync(requestToken.UserName);
-
-        if (user is null)
-            throw new BadRequestException(nameof(ErrorStrings.UserNameNotFound));
-
-        var checkPasswordResult = await _signInManager.CheckPasswordSignInAsync(user, requestToken.Password, lockoutOnFailure: true);
-
-        if (checkPasswordResult.IsLockedOut)
-            throw new BadRequestException(nameof(ErrorStrings.UserLockedOut));
-
-        if (!checkPasswordResult.Succeeded)
-            throw new BadRequestException(nameof(ErrorStrings.UserNameNotFound));
-
-        return await _jwtService.GenerateToken(user);
     }
 }
