+++ src/Tooling/Bit.Tooling.Templates/TodoTemplate/Web/Services/Implementations/TodoTemplateAuthenticationService.cs
@@ -20,7 +20,7 @@ public class TodoTemplateAuthenticationService : ITodoTemplateAuthenticationServ
 
     public async Task SignIn(SignInRequestDto dto)
     {
-        var response = await _httpClient.PostAsJsonAsync("User/SignIn", dto, ToDoTemplateJsonContext.Default.SignInRequestDto);
+        var response = await _httpClient.PostAsJsonAsync("Auth/SignIn", dto, ToDoTemplateJsonContext.Default.SignInRequestDto);
 
         var result = await response.Content.ReadFromJsonAsync(ToDoTemplateJsonContext.Default.SignInResponseDto);
